Re-style About dialog to match the java-swing family standard
Extract the About dialog out of SwingMain into its own gui/AboutDialog
class, matching every other dialog in this project and the layout
doc-scrubber/kiro-control-panel converged on: app icon west, an HTML
description panel center, and a status/close row south — replacing the
old plain vertically-stacked JLabel layout with no icon. Also escapes
the release version before it goes into the update-available HTML
label, matching the reference implementations' XSS defense against a
tampered GitHub API response.

Fixes #167
